refactor: Redondeo de reward en dto de actividad completada. #368

+++ src/main/java/trinity/play2learn/backend/activity/activity/mappers/ActivityMapper.java
@@ -51,7 +51,7 @@ public class ActivityMapper {
                                 .attempts(activity.getAttempts())
                                 .remainingAttempts(remainingAttempts)
                                 .completedAt(completedAt)
-                                .reward(reward)
+                                .reward((double) reward.intValue())
                                 .state(state)
                                 .build();
         }
